requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.8.2
mouse==0.7.1
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-ev3dev2==2.1.0.post1
six==1.16.0